refactor(LottoResults): LottoResults 클래스 위치 변경(domain -> application)
