+++ src/main/java/study/lotto/step2/application/LottoReader.java
@@ -2,7 +2,6 @@ package study.lotto.step2.application;
 
 import study.lotto.step2.domain.Lotto;
 import study.lotto.step2.domain.LottoResult;
-import study.lotto.step2.domain.LottoResults;
 
 import java.util.List;
 import java.util.stream.Collectors;
